fix: ids로 cud 시도 시 빈 배열일 경우 빈 값을 응답하도록 수정

+++ src/common/repository/common.repository.ts
@@ -29,6 +29,9 @@ export class CommonRepository<Entity extends ObjectLiteral> {
   }
 
   async findByIds(ids: string[]): Promise<Entity[]> {
+    if (ids.length === 0) {
+      return [];
+    }
     return this.repository.findBy({
       id: In(ids),
     } as unknown as FindOptionsWhere<Entity>);
@@ -78,6 +81,9 @@ export class CommonRepository<Entity extends ObjectLiteral> {
   }
 
   async softDeleteByIds(ids: string[]): Promise<DeleteResult> {
+    if (ids.length === 0) {
+      return;
+    }
     return this.repository.softDelete(ids);
   }
 
@@ -88,6 +94,9 @@ export class CommonRepository<Entity extends ObjectLiteral> {
   }
 
   async deleteByIds(ids: string[]): Promise<DeleteResult> {
+    if (ids.length === 0) {
+      return;
+    }
     return this.repository.delete(ids);
   }
 }
